update: make coords use Double

+++ app/src/main/java/com/fpf/blucon/bluetooth/BTScan.kt
@@ -2,13 +2,13 @@ package com.fpf.blucon.bluetooth
 
 data class BTScan(
     val id: Long,
-    val latitude: Float,
-    val longitude: Float,
+    val latitude: Double,
+    val longitude: Double,
     val timestamp: Long = System.currentTimeMillis(),
 )
 
 data class NewBTScan(
-    val latitude: Float,
-    val longitude: Float,
+    val latitude: Double,
+    val longitude: Double,
     val timestamp: Long = System.currentTimeMillis(),
 )
